example working for passport-hmac-auth
