+++ middleware/bos-authentication.js
@@ -57,7 +57,8 @@ function _applyCustomSecurityRequirement(app, method, route, securityReq,
                                    securityDefn, /*requiredPermissions,*/ requiredScopes) {
     //load security def middleware
     if (securityDefn['x-bos-middleware']) {
-        loader.loadConsumerModules('middleware', [path.join(global.__appDir, 'middleware', securityDefn['x-bos-middleware'])]);
+        loader.loadConsumerModules('middleware',
+            [path.join(global.__appDir, 'middleware', securityDefn['x-bos-middleware'])]);
         loader.initConsumers('middleware', [securityDefn['x-bos-middleware']], function (err) {
             if (!err) {
                 var customAuthMiddleware = loader.getConsumer('middleware', securityDefn['x-bos-middleware']);
