--- services/hmacService.js
@@ -1,4 +1,0 @@
-
-exports.getApiUser = function (apiId, done) {
-    done(null, {name: 'joe', password: 'peschi'});
-};
